chore(suite): refactor pin modal and usePinModal hook slightly
